IRGen: setup foreign calls properly
We would previously drop all calling convention information for foreign
calls on the floor when arranging a function call.  This could (would in
the case of many Windows i686 system calls) construct invalid function
calls.  We would simply mark everything as being cdecl rather than the
effective calling convention of the function call.  In the case of
windows i686, this is problematic as many functions are callee cleanup
with pass by stack, aka, `__stdcall`.  We would as a result of the
missing information pass arguments incorrectly, or in the worst case
(like on Windows i686), corrupt the stack.

The heart of this change is that when arranging the function call, if we
are calling a foreign function, we should consult clang for the extended
information for the call.  Given that SIL does not preserve this
information, we need to go back to the original foreign function
declaration and query the information.  This declaration is only
available significantly earlier in the call stack.  The newly introduced
parameters are to help identify potential gaps where we may be leaking a
calling convention attribute on the foreign call.  However, the current
state is sufficient to enable loading Foundation on Windows i686 which
uses a number of foreign calls with different conventions.