IRGen: enable the dynamic cast optimization on Windows
This accommodates lazy initialization of class metadata for the dynamic
cast elision optimization.  In the case of a lazy initialization, we
cannot assume that we are able to directly access the metadata through
a GEP, instead we should go through the type metadata accessor.
Although this is a slight bit more expensive, the hot-path of this is
roughly a function call + pointer check (if the value is in the cache,
we just return, otherwise we will have to initialize the type metadata
the first time around).

Special thanks to @atrick for the discussion, pointers, and general help
with trying to enable this optimization!