Properly canonicalize protocol compositions that include parameters.
Parameterized composition components are preserved over
non-parameterized components on the same protocol.  They are
also preserved even if they are redundant with parameterized
derived protocols.  Otherwise, they're just put in usual protocol
order.

This assumes that argument lists on the same protocol are always
redundant with each other, which may not be correct when we add
abstract constraints.  We are also, of course, still not making
an effort to diagnose conflicts.