Runtime: correct invalid C++ workaround
The placement new operator is only available when `<new>` has been
included.  Add the missing include to the header to allow us to get the
definition of the placement new allocator.  This allows us to remove the
workaround that was there for Windows, and should hopefully repair the
Android build as well.

Thanks to @grynspan for helping identify the underlying issue!

(cherry picked from commit 312f25874a7a8717422700a171f4520b009a5ce9)