Use `AbstractStorageDecl::isSettableInSwift` to prohibit direct writes to `optional` requirements
Stop pretending that an optional requirement is immutable via the `StorageImplInfo` request.
This approach has lead astray the conformance checker and might have had a negative impact
on other code paths that expect `isSettable` or `supportsMutation` to mean "has a setter".
Also, this does not work for imported foreign declarations, since they bypass the request.
Instead, use a forwarding `AbstractStorageDecl::isSettableInSwift` method that special-cases
optional requirements. This is a follow-up fix for a regression caused by #42384.