Allow specifying alternative root locations
This patch is intended to enable providing alternative toolchain
locations. This is intended primarily for cross-compilations where I
have a very specific compiler that I want to have building the runtimes
and cannot have the actual host compiler involved at any point during
the build. The provided toolchain must contain all of the necessary
binaries for the build.