Track `@main` declaration context
Add the `MainDeclContext` to the source file. This tracks the decl
context that enabled the `mainDecl` to be the actual main declaration.

In this case, the MainDeclContext will be the same as the decl-context
of main, because the main function and the `@main` are on the same
struct;

```
@main struct MainType {
  static func main() { }
}
```

In this case, the main function comes from an extension of a protocol.
The MainDecl's decl context will be the extension, but the
MainDeclContext will be the struct with the `@main` attribute
(MainType);

```
protocol Config { }
struct Config1 : Config { }

protocol Blarpy {
  associatedtype AppConfig: Config
}

extension Blarpy where AppConfig == Config {
  static func main() { print("Mega Blarpo") }
}

@main struct MainType : Blarpy {
  typealias Config = Config1
}
```

We need the original `@main` struct decl context in order to
correctly set up the substitution map in the second example.