Plumb decl context to SILGen
This patch plumbs the FileUnit, and with it, the correct decl context
down to the SILGen layer. Note that this does not affect the application
main, only the function-decl main.

At the moment, this is NFC since the decl-context is that of `$main`,
which will always be the declcontext with the `@main` attribute. Once
`$main` is removed, they may not match. e.g. when the `@main` is on a
struct that conforms to a protocol with an extension containing the main
function implementation.