Fix StringObject.nativeStorage to avoid bitcasting to a reference.
Use an Unmanaged reference when reinterpreting a strong reference to
any unmanaged (trivial) representation.

Mutating StringGuts methods, such as appendInPlace, are currently
miscompiled just by reordering the optimization pipeline. The fact
that 'self' is guaranteed within a method scope only applies to
nonmutating methods. In mutating methods, the contents of 'self' are
conceptually destroyed at the beginning of the method. Futhermore, the
compiler cannot provide any lifetime guarantees in the presence of a
bitcasting from an integer to a strong reference.

As a basic rule, Swift code should never directly bitcast an integer
to a reference. There's no reasonable way for the compiler to support
this. Generating a strong reference out of thin air requires an
explicit conversion from an unmanaged reference.

This rule will need to be clearly documented. To start with, it
will be explained in an upcoming proposal for object lifetimes.